Rename physical processor to clickhouse processor
