+++ snuba/datasets/plans/query_plan.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from typing import Callable, Sequence
 
 from snuba.clickhouse.query import Query
-from snuba.query.processors.physical import QueryProcessor
+from snuba.clickhouse.processors import QueryProcessor
 from snuba.request import Request
 from snuba.request.request_settings import RequestSettings
 from snuba.web import QueryResult
