+++ snuba/query/processors/events_column_processor.py
@@ -1,6 +1,6 @@
 from snuba.clickhouse.query import Query
 from snuba.query.expressions import Column, Expression, FunctionCall, Literal
-from snuba.query.processors.physical import QueryProcessor
+from snuba.clickhouse.processors import QueryProcessor
 from snuba.request.request_settings import RequestSettings
