+++ snuba/query/processors/prewhere.py
@@ -2,7 +2,7 @@ from typing import Optional, Sequence
 
 from snuba import settings, util
 from snuba.clickhouse.query import Query
-from snuba.query.processors.physical import QueryProcessor
+from snuba.clickhouse.processors import QueryProcessor
 from snuba.query.types import Condition
 from snuba.request.request_settings import RequestSettings
